finish assignment

+++ programmingInPython/week1/exercise2.py
@@ -9,13 +9,13 @@
 #   Your total bill is $ 9.74
 
 # Modify the line below
-coffee = int(input('1 coffee @: $ '))
+coffee = float(input('1 coffee @: $ '))
 
 # Modify the line below
-sandwich = int(input('1 sandwich @: $ '))
+sandwich = float(input('1 sandwich @: $ '))
 
 # Modify the line below
-cake = int(input('1 cake @: $ '))
+cake = float(input('1 cake @: $ '))
 
 bill_total = coffee + sandwich + cake
